Type::TypeId to enum class changes

+++ src/traffic_cop/traffic_cop.cpp
@@ -493,7 +493,8 @@ FieldInfo TrafficCop::GetColumnFieldForValueType(std::string column_name,
     default: {
       // Type not Identified
       LOG_ERROR("Unrecognized field type '%s' [%d] for field '%s'",
-                TypeIdToString(column_type).c_str(), column_type,
+                TypeIdToString(column_type).c_str(),
+                static_cast<int>(column_type),
                 column_name.c_str());
       field_type = PostgresValueType::TEXT;
       field_size = 255;
